Add dependencies for local server
Add the required dependencies for the local server to be implemented in the future. The AutoTuneServlet and CSVHandler class has also been added but is not yet implemented. The build.gradle has also been changed to use variables as the versions for dependencies to allow for version bumps to be easier.