Make CsvHandler an async event
Change the CSV handler to a new async event. This is beneficial because it means it can be called asynchronously which improves performance. The CsvWrite event can become expensive when many items are present and many time periods have passed. This allows it to not reduce server performance and be called more frequently. It is now called at each price update.