Implement embedded web server
Re-implemented the embedded web server using jetty. The web server files are now located in resources and created in the Auto-Tune data folder under /web. Price history data is written to web/data/data.csv using the CsvWriteEvent. The LocalServer class stores the server startup information and commands. The web server contains a search box to search for items in the database where their prices will be displayed as a graph on the website. The source code for the integrated web server will be available soon .