Fix formatting
Fixed formatting on Config class. (We use the google checkstyle https://checkstyle.sourceforge.io/google_style.html). The version was de-bumped because it can stay on 0.14.0 pre-release until 0.14 is released. Removed the UtilFunctions class as the function to get the player count can just be done in the TimePeriodEvent class. The enable-sell-limits configuration option should remain true by default because removing it on smaller servers can cause exploit issues.