Format code to remove all exceptions in google java checkstyle
Set tab indent size to 4, and removed "AvoideStartImport" module. Fixed the code so that it conformed to the rest of the options in google's java checktyle. Google requires imports to be in alphabetical order with no gaps between as well as requiring a max line length of 100. These were the main changes I had to make.

Also made sure 4 spaces were used as a tab in the codebase as much as possible.